[K/N] Widen thread_id to uintptr_t
Currently inside the K/N runtime thread id is assumed to be 32-bit.

But in practice it's not always the case. So, when thread id provided
by `pthread_threadid_np` on iOS doesn't fit into 32 bits, runtime 
crashes. We've seen this in practice.

This commit makes thread id pointer size instead, and adjusts the
related code in runtime and compiler. Having it pointer size is
practically enough to guarantee that atomic operations on it will
still be possible.

^KT-70365

+++ kotlin-native/runtime/src/main/cpp/Runtime.h
@@ -6,6 +6,7 @@
 #ifndef RUNTIME_RUNTIME_H
 #define RUNTIME_RUNTIME_H
 
+#include <cstdint>
 #include "Porting.h"
 #include "Memory.h"
 
@@ -32,8 +33,8 @@ void Kotlin_shutdownRuntime();
 // Appends given node to an initializer list.
 RUNTIME_NOTHROW void AppendToInitializersTail(struct InitNode*);
 
-void CallInitGlobalPossiblyLock(int* state, void (*init)());
-void CallInitThreadLocal(int volatile* globalState, int* localState, void (*init)());
+void CallInitGlobalPossiblyLock(uintptr_t* state, void (*init)());
+void CallInitThreadLocal(uintptr_t volatile* globalState, uintptr_t* localState, void (*init)());
 
 bool Kotlin_memoryLeakCheckerEnabled();
